add package file structure
