+++ src/python/datapreprocessor/test/csvreader_tests.py
@@ -1,0 +1,6 @@
+import unittest
+import datapreprocessor
+
+class CsvFileReaderTest(unittest.TestCase):
+    def test(self):
+        self.assertEqual(1, 2)
